interim refactor of towards HashSetManager

+++ ImageGallery/src/org/sleuthkit/autopsy/imagegallery/gui/navpanel/TreeNodeComparators.java
@@ -28,31 +28,34 @@ import javafx.scene.control.TreeItem;
 enum TreeNodeComparators implements Comparator<TreeItem<TreeNode>>, NonNullCompareable {
 
     ALPHABETICAL("Group Name") {
-        @Override
+                @Override
         public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
 
-            return o1.getValue().getGroup().groupKey.getValue().toString().compareTo(o2.getValue().getGroup().groupKey.getValue().toString());
-        }
-    }, HIT_COUNT("Hit Count") {
-        @Override
-        public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
+                    return o1.getValue().getGroup().groupKey.getValue().toString().compareTo(o2.getValue().getGroup().groupKey.getValue().toString());
+                }
+            },
+    HIT_COUNT("Hit Count") {
+                @Override
+                public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
 
-            return -Integer.compare(o1.getValue().getGroup().getHashSetHitsCount(), o2.getValue().getGroup().getHashSetHitsCount());
-        }
-    }, FILE_COUNT("Group Size") {
-        @Override
-        public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
+                    return -Long.compare(o1.getValue().getGroup().getHashSetHitsCount(), o2.getValue().getGroup().getHashSetHitsCount());
+                }
+            },
+    FILE_COUNT("Group Size") {
+                @Override
+                public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
 
-            return -Integer.compare(o1.getValue().getGroup().getSize(), o2.getValue().getGroup().getSize());
-        }
-    }, HIT_FILE_RATIO("Hit Density") {
-        @Override
-        public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
+                    return -Integer.compare(o1.getValue().getGroup().getSize(), o2.getValue().getGroup().getSize());
+                }
+            },
+    HIT_FILE_RATIO("Hit Density") {
+                @Override
+                public int nonNullCompare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
 
-            return -Double.compare(o1.getValue().getGroup().getHashSetHitsCount() / (double) o1.getValue().getGroup().getSize(),
-                    o2.getValue().getGroup().getHashSetHitsCount() / (double) o2.getValue().getGroup().getSize());
-        }
-    };
+                    return -Double.compare(o1.getValue().getGroup().getHashSetHitsCount() / (double) o1.getValue().getGroup().getSize(),
+                            o2.getValue().getGroup().getHashSetHitsCount() / (double) o2.getValue().getGroup().getSize());
+                }
+            };
 
     @Override
     public int compare(TreeItem<TreeNode> o1, TreeItem<TreeNode> o2) {
